적과 총알 충돌시 체력감소 및 죽음 구현

+++ Assets/Scripts/EnemyData.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "Scriptable/EnemyData", fileName = "Enemy Data")]
 public class EnemyData : ScriptableObject {
 	public string type = "Normal";
-	public int health = 100;
+	public float health = 100;
 	public int sp = 10;
 	public float speed = 1f;
 	public Sprite sprite;
